Rename file contains `promisifySend` function
